Change the status of items to Done
Add a "complete" button to each item in the list of items.
When the button is clicked the item moves to the Done list.
The item is not displayed. (The Done list is not displayed.)
TODO It would be better to see both lists, To Do and Done.
TODO To Do items are in the wrong Trello list, Doing instead of To Do.